Add puzzle parsing and integration to game scene
Introduced a new puzzle class with string parsing to represent battery configurations. Updated the battery class to support string serialization and deserialization. Integrated puzzle setup into the game scene, replacing hardcoded test setup with dynamic puzzle loading. Refactored battery display and toggle logic to use size_t for indices and improved type safety.

+++ src/energy/components/battery_display.cpp
@@ -61,7 +61,7 @@ auto battery_display::update(const float delta) -> engine::result<> {
 	}
 
 	for(size_t i = 0; i < segments_.size(); ++i) {
-		const auto color_index = battery_->get().at(static_cast<int>(i));
+		const auto color_index = battery_->get().at(i);
 		const auto color = energy_colors.at(color_index);
 		segments_.at(i).set_tint(color);
 	}
